requirements/base.txt
django==2.1.5
django-environ==0.4.5
psycopg2-binary==2.9.1
Twisted==19.7.0
celery==4.4.7
django-crispy-forms==1.7.0
crispy-forms-materialize==0.2
django_celery_results==1.0.1
channels==2.2.0
djangorestframework==3.7.7
django-braces==1.12.0
django_celery_beat==1.1.0
django-phonenumber-field==2.0.0
django-cleanup==2.1.0
social-auth-app-django==2.1.0
django-datatables-view==1.14.0
django-mail-templated==2.6.5
openpyxl==3.0.3
Pillow==8.3.1
django-storages==1.7.1
django-money==0.12.3
PyPDF2==1.26.0
XlsxWriter==1.0.2
python-docx==0.8.6
pdfkit==0.6.1
diff-match-patch==20121119
cssutils==1.0.2
lxml==4.6.3
bs4==0.0.1
django-autocomplete-light==3.3.0rc6
django-mathfilters==0.4.0
channels-redis==2.4.0
num2words==0.5.7
html2text==2018.1.9
PyPDF2==1.26.0
redis==3.2.1
reportlab==3.5.23
flower==0.9.3
daphne==2.3.0
azure==4.0.0
whitenoise==4.1.2
sentry-sdk==0.14.1
